Starting over from scratch
I fill the need to start over as a lot of the code start to get junked
together and was falling apart. From here on out i'm going to rewrite
what i had using minimal code from the old mods. This way i get a
cleaner and more reusable code for the mod